move formatUser function to helper folder
