+++ src/app/modules/item/components/item-chapter-view/group-situation-chapter-view/group-situation-chapter-view.component.ts
@@ -4,9 +4,9 @@ import { map, switchMap } from 'rxjs/operators';
 import { Group } from 'src/app/core/components/group-nav-tree/group';
 import { GetGroupChildrenService } from 'src/app/modules/group/http-services/get-group-children.service';
 import { fetchingState, isReady, readyState } from 'src/app/shared/helpers/state';
+import { formatUser } from 'src/app/shared/helpers/user';
 import { GetGroupDescendantsService } from 'src/app/shared/http-services/get-group-descendants.service';
 import { GetGroupProgressService } from 'src/app/shared/http-services/get-group-progress.service';
-import { formatUser } from 'src/app/shared/pipes/userDisplay';
 import { GetItemChildrenService, ItemChild } from '../../../http-services/get-item-children.service';
 import { ItemData } from '../../../services/item-datasource.service';
 import { TypeFilter } from './composition-filter/composition-filter.component';
